outlook/requirements.txt
fastmcp>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
msal>=1.28.0
